Restructure ykbuild a bit.
Separate out the clangd compile database generation stuff from other
misc stuff.
